shifted browser delay timer to UIBrowserBridge
