+++ src/ui/issuepanel/IssuePanel.java
@@ -3,7 +3,6 @@ package ui.issuepanel;
 import java.lang.ref.WeakReference;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.concurrent.TimeUnit;
 
 import javafx.event.EventHandler;
 import javafx.scene.control.ListCell;
@@ -18,7 +17,6 @@ import javafx.util.Callback;
 import model.Model;
 import model.TurboIssue;
 import service.ServiceManager;
-import service.TickingTimer;
 import ui.UI;
 import ui.components.NavigableListView;
 import ui.issuecolumn.ColumnControl;
@@ -40,11 +38,6 @@ public class IssuePanel extends IssueColumn {
 	private final KeyCombination defaultSizeWindow = new KeyCodeCombination(KeyCode.D, KeyCombination.CONTROL_DOWN);
 	private HashMap<Integer, Integer> issueCommentCounts = new HashMap<>();
 
-	private static final int BROWSER_REQUEST_DELAY = 500; //milliseconds
-	private static TickingTimer timer;
-	private static int nextIssueId;
-	private static int nextColumnIndex;
-
 	public IssuePanel(UI ui, Stage mainStage, Model model, ColumnControl parentColumnControl, int columnIndex, TurboCommandExecutor dragAndDropExecutor) {
 		super(ui, mainStage, model, parentColumnControl, columnIndex, dragAndDropExecutor);
 		this.model = model;
@@ -53,12 +46,6 @@ public class IssuePanel extends IssueColumn {
 		listView = new NavigableListView<>();
 		setupListView();
 		getChildren().add(listView);
-		if (timer == null) {
-			nextIssueId = -1;
-			nextColumnIndex = -1;
-			timer = createTickingTimer();
-			timer.start();
-		}
 
 		refreshItems();
 	}
@@ -135,29 +122,12 @@ public class IssuePanel extends IssueColumn {
 		setupKeyboardShortcuts();
 		listView.setOnItemSelected(i -> {
 			TurboIssue issue = listView.getItems().get(i);
-			nextIssueId = issue.getId();
-			nextColumnIndex = columnIndex;
-			timer.restart();
-			if (timer.isPaused()) {
-				timer.resume();
-			}
+			ui.triggerEvent(new IssueSelectedEvent(issue.getId(), columnIndex));
 			if (issueHasNewComments(issue)) {
 				issueCommentCounts.put(issue.getId(), issue.getCommentCount());
 				refreshItems();
 			}
 		});
-	}
-
-	private TickingTimer createTickingTimer() {
-		return new TickingTimer("Browser Request Delay Timer", BROWSER_REQUEST_DELAY, integer -> {
-            // do nothing for each tick
-        }, () -> {
-			if (nextIssueId >= 0) {
-				System.out.println("navigating to issue #" + nextIssueId + " from column: " + nextColumnIndex);
-				ui.triggerEvent(new IssueSelectedEvent(nextIssueId, nextColumnIndex));
-			}
-			timer.pause();
-		}, TimeUnit.MILLISECONDS);
 	}
 	
 	private void setupKeyboardShortcuts(){
